runclient long fix etc
git-svn-id: http://astropew.googlecode.com/svn/trunk@78 d889380e-1d37-11de-8b5c-13accb87f508

+++ src/common/Util.java
@@ -66,7 +66,7 @@ public class Util {
 	 */
 	public static long getLong(byte[] b, int off) {
 		long val = 0;
-		int mult = 1;
+		long mult = 1;
 		for (int i=off+7; i>=off; i--){
 				int toadd = b[i];
 				if ( toadd < 0){
@@ -91,7 +91,7 @@ public class Util {
 		dst[offset+0] = (byte)( (i >> 24) % 256 );
 		dst[offset+1] = (byte)( (i >> 16) % 256 );
 		dst[offset+2] = (byte)( (i >>  8) % 256 );
-		dst[offset+3] = (byte)( (i	  ) % 256 );
+		dst[offset+3] = (byte)( (i      ) % 256 );
 	}
 	
 	/**
@@ -132,7 +132,7 @@ public class Util {
 		dst[offset+4] = (byte)( (l >> 24) % 256 );
 		dst[offset+5] = (byte)( (l >> 16) % 256 );
 		dst[offset+6] = (byte)( (l >>  8) % 256 );
-		dst[offset+7] = (byte)( (l	  ) % 256 );
+		dst[offset+7] = (byte)( (l      ) % 256 );
 	}
 
 	/**
